// frontend/src/api/model/getTrendingMoviesPeriod.ts
/**
 * Generated by orval v8.4.0 🍺
 * Do not edit manually.
 * Movie Review System API
 * Backend API for Movie Review Website
 * OpenAPI spec version: v1.0
 */

export type GetTrendingMoviesPeriod =
  (typeof GetTrendingMoviesPeriod)[keyof typeof GetTrendingMoviesPeriod];

export const GetTrendingMoviesPeriod = {
  DAILY: "DAILY",
  WEEKLY: "WEEKLY",
  MONTHLY: "MONTHLY",
} as const;